Remodeled window postion's internal representation
Instead of storing position only in the window and iterating
four times over all windows filtering by dock,
there is one vector for each dock storing the windows' ids to be
rendered.
This enables implementing sorting inside a dock later as well as
unifies the handling of the side docks and the center dock,
which is a stack now.